Deleted SessionCookieUtil.scala and its test file, moved code into EnterClientsUTRController
