+++ app/controllers/agent/ConfirmClientUTRController.scala
@@ -23,11 +23,14 @@ import config.featureswitch.FeatureSwitching
 import config.{AgentItvcErrorHandler, FrontendAppConfig}
 import controllers.agent.sessionUtils.SessionKeys
 import models.sessionData.SessionCookieData
+import models.sessionData.SessionDataPostResponse.{SessionDataPostFailure, SessionDataPostSuccess}
+import play.api.Logger
 import play.api.i18n.I18nSupport
-import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
+import play.api.mvc.{Action, AnyContent, MessagesControllerComponents, Request, Result}
 import services.SessionDataService
+import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-import utils.{AuthenticatorPredicate, SessionCookieUtil}
+import utils.AuthenticatorPredicate
 import views.html.agent.confirmClient
 
 import javax.inject.{Inject, Singleton}
@@ -43,7 +46,7 @@ class ConfirmClientUTRController @Inject()(confirmClient: confirmClient,
                                            val appConfig: FrontendAppConfig,
                                            val itvcErrorHandler: AgentItvcErrorHandler,
                                            val ec: ExecutionContext)
-  extends FrontendController(mcc) with FeatureSwitching with I18nSupport with SessionCookieUtil {
+  extends FrontendController(mcc) with FeatureSwitching with I18nSupport {
 
   def show: Action[AnyContent] = authActions.asMTDAgentWithUnconfirmedClient { implicit user =>
     Ok(confirmClient(
@@ -78,5 +81,24 @@ class ConfirmClientUTRController @Inject()(confirmClient: confirmClient,
   }
 
   lazy val backUrl: String = controllers.agent.routes.EnterClientsUTRController.show.url
+
+  def getSessionDataStorageFS: Boolean = appConfig.isSessionDataStorageEnabled
+
+  def handleSessionCookies(sessionCookieData: SessionCookieData)(codeBlock: Seq[(String, String)] => Future[Result])
+                          (implicit hc: HeaderCarrier, request: Request[_], ec: ExecutionContext): Future[Result] = {
+    if (getSessionDataStorageFS) {
+      sessionDataService.postSessionData(sessionCookieData.toSessionDataModel).flatMap {
+        case Left(value: SessionDataPostFailure) =>
+          Logger("application").error(s"[Agent] Posting user session data was unsuccessful. Status: ${value.status}, error message: ${value.errorMessage}")
+          Future.successful(itvcErrorHandler.showInternalServerError())
+        case Right(value: SessionDataPostSuccess) =>
+          Logger("application").info(s"[Agent] Posting user session data was successful. Status: ${value.status}")
+          codeBlock(sessionCookieData.toSessionCookieSeq)
+      }
+    } else {
+      Logger("application").info(s"[Agent] GetUserSessionApi feature switch was off so session data has not been posted to the session-data service")
+      codeBlock(sessionCookieData.toSessionCookieSeq)
+    }
+  }
 
 }
